product-reviews

+++ isd-farmers-market/src/main/java/com/example/isdfarmersmarket/dao/models/ProductReview.java
@@ -20,7 +20,7 @@ public class ProductReview {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "hibernate_sequence")
     private Long id;
     private String content;
-    private float rating;
+    private Integer rating;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "creator_id")
